+++ front/src/assets/styles/_iarbre.sass
@@ -12,6 +12,7 @@ $dark-green: #32412d
 $light-green: #92a48d
 $brown: #32312d
 $white: #ffffff
+$red: red
 
 $main-font: "Bricolage Grotesque"
 $accent-font: "Sligoil"
